pyproject.toml
[project]
name = "trae-account-creator"
version = "0.1.0"
description = "Trae Account Creator with Gmail IMAP and Cloudflare"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "playwright",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.11.1",
]
